Extract prefetch hash helpers into shared utility
