+++ packages/docs/src/data/articles.ts
@@ -5105,15 +5105,6 @@ export const articles = [
 		slug: 'prefetch',
 	},
 	{
-		id: 'resumable-prefetch',
-		title: 'resumablePrefetch()',
-		relativePath: 'docs/resumable-prefetch.mdx',
-		compId: 'articles-docs-resumable-prefetch',
-		crumb: 'API',
-		noAi: false,
-		slug: 'resumable-prefetch',
-	},
-	{
 		id: 'preload-audio',
 		title: 'preloadAudio()',
 		relativePath: 'docs/preload/preload-audio.mdx',
@@ -5744,6 +5735,15 @@ export const articles = [
 		crumb: 'Ecosystem',
 		noAi: false,
 		slug: 'resources',
+	},
+	{
+		id: 'resumable-prefetch',
+		title: 'resumablePrefetch()',
+		relativePath: 'docs/resumable-prefetch.mdx',
+		compId: 'articles-docs-resumable-prefetch',
+		crumb: 'API',
+		noAi: false,
+		slug: 'resumable-prefetch',
 	},
 	{
 		id: 'rive/index',
